+++ integration/app/counter/counter.component.ts
@@ -1,3 +1,4 @@
+import { CommonModule } from '@angular/common';
 import { Component, OnInit, Signal } from '@angular/core';
 import { Store } from '@ngxs/store';
 import { Observable } from 'rxjs';
@@ -7,7 +8,8 @@ import { CounterState, CounterStateModel } from '@integration/counter/counter.st
 
 @Component({
   selector: 'counter',
-  templateUrl: './counter.component.html'
+  templateUrl: './counter.component.html',
+  imports: [CommonModule]
 })
 export class CounterComponent implements OnInit {
   counter$: Observable<CounterStateModel> = this._store.select(CounterState.getCounterState);