+++ integration/app/list/list.component.ts
@@ -1,3 +1,4 @@
+import { CommonModule } from '@angular/common';
 import { Component, Signal } from '@angular/core';
 import { RouterStateSnapshot } from '@angular/router';
 import { Store } from '@ngxs/store';
@@ -8,7 +9,8 @@ import { ListState } from '@integration/list/list.state';
 
 @Component({
   selector: 'app-list',
-  templateUrl: './list.component.html'
+  templateUrl: './list.component.html',
+  imports: [CommonModule]
 })
 export class ListComponent {
   list$: Observable<string[]> = this._store.select(ListState.getListState);