+++ packages/router-plugin/src/router.state.ts
@@ -1,4 +1,4 @@
-import { NgZone, Injectable, OnDestroy, Injector } from '@angular/core';
+import { NgZone, Injectable, OnDestroy, inject } from '@angular/core';
 import {
   NavigationCancel,
   NavigationError,
@@ -10,10 +10,9 @@ import {
   NavigationEnd,
   Event
 } from '@angular/router';
-import { Action, createSelector, State, StateContext, StateToken, Store } from '@ngxs/store';
+import { Action, Selector, State, StateContext, StateToken, Store } from '@ngxs/store';
 import {
   NavigationActionTiming,
-  NgxsRouterPluginOptions,
   ɵNGXS_ROUTER_PLUGIN_OPTIONS
 } from '@ngxs/router-plugin/internals';
 import { ReplaySubject } from 'rxjs';
@@ -59,6 +58,12 @@ export const ROUTER_STATE_TOKEN = new StateToken<RouterStateModel>('router');
 })
 @Injectable()
 export class RouterState implements OnDestroy {
+  private _store = inject(Store);
+  private _router = inject(Router);
+  private _serializer: RouterStateSerializer<RouterStateSnapshot> =
+    inject(RouterStateSerializer);
+  private _ngZone = inject(NgZone);
+
   /**
    * Determines how navigation was performed by the `RouterState` itself
    * or outside via `new Navigate(...)`
@@ -77,34 +82,23 @@ export class RouterState implements OnDestroy {
 
   private _lastEvent: Event | null = null;
 
-  private _options: NgxsRouterPluginOptions | null = null;
+  private _options = inject(ɵNGXS_ROUTER_PLUGIN_OPTIONS);
 
   private _destroy$ = new ReplaySubject<void>(1);
 
-  static state = /* @__PURE__ */ createSelector(
-    [ROUTER_STATE_TOKEN],
-    (state: RouterStateModel<RouterStateSnapshot>) => {
-      // The `state` is optional if the selector is invoked before the router
-      // state is registered in NGXS.
-      return state?.state;
-    }
-  );
+  @Selector()
+  static state<T = RouterStateSnapshot>(state: RouterStateModel<T>) {
+    // The `state` is optional if the selector is invoked before the router
+    // state is registered in NGXS.
+    return state?.state;
+  }
 
-  static url = /* @__PURE__ */ createSelector(
-    [ROUTER_STATE_TOKEN],
-    state => state?.state?.url
-  );
+  @Selector()
+  static url(state: RouterStateModel): string | undefined {
+    return state?.state?.url;
+  }
 
-  constructor(
-    private _store: Store,
-    private _router: Router,
-    private _serializer: RouterStateSerializer<RouterStateSnapshot>,
-    private _ngZone: NgZone,
-    injector: Injector
-  ) {
-    // Note: do not use `@Inject` since it fails on lower versions of Angular with Jest
-    // integration, it cannot resolve the token provider.
-    this._options = injector.get(ɵNGXS_ROUTER_PLUGIN_OPTIONS, null);
+  constructor() {
     this._setUpStoreListener();
     this._setUpRouterEventsListener();
   }