+++ packages/router-plugin/tests/issues/issue-1718-url-recognition.spec.ts
@@ -20,19 +20,22 @@ import { RouterState, withNgxsRouterPlugin } from '../..';
 describe('URL recognition in guards (https://github.com/ngxs/store/issues/1718)', () => {
   @Component({
     selector: 'app-root',
-    template: '<router-outlet></router-outlet>'
+    template: '<router-outlet></router-outlet>',
+    standalone: false
   })
   class RootComponent {}
 
   @Component({
     selector: 'app-home',
-    template: '<a class="navigate-to-details" routerLink="/details">Details</a>'
+    template: '<a class="navigate-to-details" routerLink="/details">Details</a>',
+    standalone: false
   })
   class HomeComponent {}
 
   @Component({
     selector: 'app-details',
-    template: 'Details page'
+    template: 'Details page',
+    standalone: false
   })
   class DetailsComponent {}