+++ packages/store/src/internal/dispatcher.ts
@@ -1,11 +1,10 @@
-import { Injectable, NgZone } from '@angular/core';
-import { EMPTY, forkJoin, Observable, of, Subject, throwError } from 'rxjs';
-import { exhaustMap, filter, map, shareReplay, take } from 'rxjs/operators';
+import { inject, Injectable, Injector, NgZone, runInInjectionContext } from '@angular/core';
+import { forkJoin, Observable, of, Subject, throwError } from 'rxjs';
+import { filter, map, mergeMap, shareReplay, take } from 'rxjs/operators';
 
 import { getActionTypeFromInstance } from '@ngxs/store/plugins';
 import { ɵPlainObject, ɵStateStream } from '@ngxs/store/internals';
 
-import { compose } from '../utils/compose';
 import { ActionContext, ActionStatus, InternalActions } from '../actions-stream';
 import { PluginManager } from '../plugin-manager';
 import { InternalNgxsExecutionStrategy } from '../execution/internal-ngxs-execution-strategy';
@@ -23,14 +22,13 @@ export class InternalDispatchedActionResults extends Subject<ActionContext> {}
 
 @Injectable({ providedIn: 'root' })
 export class InternalDispatcher {
-  constructor(
-    private _ngZone: NgZone,
-    private _actions: InternalActions,
-    private _actionResults: InternalDispatchedActionResults,
-    private _pluginManager: PluginManager,
-    private _stateStream: ɵStateStream,
-    private _ngxsExecutionStrategy: InternalNgxsExecutionStrategy
-  ) {}
+  private _ngZone = inject(NgZone);
+  private _actions = inject(InternalActions);
+  private _actionResults = inject(InternalDispatchedActionResults);
+  private _pluginManager = inject(PluginManager);
+  private _stateStream = inject(ɵStateStream);
+  private _ngxsExecutionStrategy = inject(InternalNgxsExecutionStrategy);
+  private _injector = inject(Injector);
 
   /**
    * Dispatches event(s).
@@ -72,7 +70,7 @@ export class InternalDispatcher {
     const prevState = this._stateStream.getValue();
     const plugins = this._pluginManager.plugins;
 
-    return compose([
+    return compose(this._injector, [
       ...plugins,
       (nextState: any, nextAction: any) => {
         if (nextState !== prevState) {
@@ -99,21 +97,52 @@ export class InternalDispatcher {
   private createDispatchObservable(
     actionResult$: Observable<ActionContext>
   ): Observable<ɵPlainObject> {
-    return actionResult$
-      .pipe(
-        exhaustMap((ctx: ActionContext) => {
-          switch (ctx.status) {
-            case ActionStatus.Successful:
-              // The `createDispatchObservable` function should return the
-              // state, as its result is utilized by plugins.
-              return of(this._stateStream.getValue());
-            case ActionStatus.Errored:
-              return throwError(() => ctx.error);
-            default:
-              return EMPTY;
-          }
-        })
-      )
-      .pipe(shareReplay());
+    return actionResult$.pipe(
+      mergeMap((ctx: ActionContext) => {
+        switch (ctx.status) {
+          case ActionStatus.Successful:
+            // The `createDispatchObservable` function should return the
+            // state, as its result is used by plugins.
+            return of(this._stateStream.getValue());
+          case ActionStatus.Errored:
+            return throwError(() => ctx.error);
+          default:
+            // Once dispatched or canceled, we complete it immediately because
+            // `dispatch()` should emit (or error, or complete) as soon as it succeeds or fails.
+            return of();
+        }
+      }),
+      shareReplay()
+    );
   }
 }
+
+type StateFn = (...args: any[]) => any;
+
+/**
+ * Composes a array of functions from left to right. Example:
+ *
+ *      compose([fn, final])(state, action);
+ *
+ * then the funcs have a signature like:
+ *
+ *      function fn (state, action, next) {
+ *          console.log('here', state, action, next);
+ *          return next(state, action);
+ *      }
+ *
+ *      function final (state, action) {
+ *          console.log('here', state, action);
+ *          return state;
+ *      }
+ *
+ * the last function should not call `next`.
+ */
+const compose =
+  (injector: Injector, funcs: StateFn[]) =>
+  (...args: any[]) => {
+    const curr = funcs.shift()!;
+    return runInInjectionContext(injector, () =>
+      curr(...args, (...nextArgs: any[]) => compose(injector, funcs)(...nextArgs))
+    );
+  };