+++ packages/store/tests/state.spec.ts
@@ -275,7 +275,8 @@ describe('State', () => {
 
     @Component({
       selector: 'app-root',
-      template: ''
+      template: '',
+      standalone: false
     })
     class MockComponent implements OnInit, AfterViewInit {
       public ngOnInit(): void {